Parse the ASN.1 declaration once, not once per field.
There is no reason to build a new ASN.1 parser for each field in the CSR.

Cache the Convert::ASN1 object in the Crypt::PKCS10 object, and simply find() the desired macro for each field.

This could actually be done once when Crypt::PKCS10 loads, but this approach leaves ope the option of registering unique ASN.1 for each Crypt::PKCS10.  (Not likely to happen.)