Add more OIDs. Generate OID doc automagically. Support API with standard names
Added OIDs for more DN components & encryption.

Added extensions() method to return the names of all extensions in the CSR.

subjectAltName can now return the list of object names contained in the value, and can return the entire name as a string.  It still returns the value(s) of a given object type if one is specified.

Added internal code to generate POD table of OIDs and names that are recognized.

Deprecate OID names with spaces and non-RFC capitalization.

Provide class method to set API version:
  Version 0 uses the old names. (Unwise and deprecated)
  Version 1 uses the new names.

The default version is 0 for backward compatibility.  It may change to 1 later.

Tests now use version 1.  Add tests for subjectAltName.

Improved documentation, part 1.