Provide useful attribute information with API v1, add GN/SN abbrevs in DNs
attributes method provided a useless view of ASN.1 in API v0.

For APIv1, change the attribute method to return the names of
attributes present in the CSR.  Typically these are challengePassword and/or unstructuredName.  Also, retrieving the value of an attribute by name.

At the moment, only string values are supported.

The APIv0 version of attributes() never worked as documented and exposed barely-parses ASN.1.  (The doc said it returned a hashref; it returned a hash.  The doc said "run Data::Dumper on what's returned to figure out what it is".  This is still available with APIv0 - but DEPRECATED.