Fixes for README generation
Automagic generation of the README using Pod::Weaver requires some handstands.

This is closer, but still has some rough edges due to issues with the tools.

At this point, all the AUTHOR, COPYRIGHT, and BUG paragraphs are generated by Pod::Weaver from the dist.ini.

There are some redundant lines in the doc to make the tools happy, and some ugliness in how the build works.