Optionally return all available curve detail from ECC keys using subjectPublicKeyParams(1)
Since it's available, might as well return extended info as delivered by Crypt::PK::ECC.
The detail items are all constants for a given curve (OID), so it doesn't seem worthwhile to
copy the data unless requested.  Plus it would give the impression that the data varies.

Some items are duplicated in the standard return.  Also, note that the curve_name in
the detail may not match the standard curve name.  The detail is provided by another
module.  It varies by curve, and the item names are controlled by that module.

Improve the documentation for subjectPublicKeyParams.