Verify CSR signature in new() by default.  Provide checkSignature().
All CSRs should have valid signatures.  Verify them in new() by default,
as this is the safest action.  It also simplifies callers who wish to verify,
as the code is non-trivial in the general case and depends on the
signature type.

The verifySignature option allows the caller to suppress this if
it knows better.

Update the tests to include CSRs with bad signatures.

Update the utilities in examples/ to reflect these changes.