Allow new() to open a  file by name & accept options in a hashref.
readFile => 1 will treat the $der parameter as a filename, which will
be opened, read, and closed.

This saves the caller from calling open() in a common case.

This is restricted to a filename.  If you want to pass a socket or a
pipe, use a tempfile or flock(), use a file handle without readFile.