Address dependency on Crypt::PK::ECC versions, add getAPIversion, as_string & fix length decode
Crypt::PK::ECC doesn't have a version, but older versions do cause test failures.
CPANTS found this.

Crypt::PK::ECC is distributed with CryptX ,so add a dependency on it.
Also, test for some of the known failures and BAIL_OUT with a more informative message.

Add getAPIversion in case someone needs to save/restore.

Change stringify operator to as_string, which is overloadable.  Can be called, but
this is primarly so that any superclass can extend the output.

Correct DER parsing of expected length to allow any length.