Encode PEM with max line length of 64. Fix test failures on some editions of Perl 5.8.8
For compatibility with strict parsers, encode PEM with a line length of 64 as
specified in RFC7468.  MIME::Base64 uses 76, which is the MIME line length.

Test failures are due to Perl bug 39185, which is encountered by 02_base.t.

Sometimes qr//ms doesn't match when it should.  qr/(?ms:)/ is happy.

Some distributions of 5.8.8 have this fixed...