Make support for signature verification and ECC keys optional, fix APIv0 stringify and add new() options
DSA isn't universally popular.  ECC is omitted from some early OS distributions
due to IP concerns.

Install warns if at least one of DSA, ECC or RSA is not installed.

Note that signature verification is on by default; it's intentionally difficult to run without it.

Decoding ECC public keys requires ECC.

Tests are skipped if they rely on the external modules.

If any support module is installed, it must pass a simple test in Makefile.PL to verify OpenSSL support.

subjectPublicKeyParams and checkSignature set $error if a CSR requires a module or hash that isn't present/supported.

Clean up error handling.  Provide instance method for error

Fix stringify's interactions with API v0.

Add PEMonly and binaryMode options.

Add a few tests for API v0.

Also, fix some of the oddities that are caused by an API version switch. (Which still is not supported.)
The toughest issue is that the legacy structure stores OID names, which differ with version.

In the long-term, API v0 should be removed...